+++ test/OpenSpaceToolkit/Astrodynamics/GuidanceLaw/QLaw.test.cpp
@@ -153,93 +153,3 @@ TEST_F(OpenSpaceToolkit_Astrodynamics_Dynamics_Thruster_GuidanceLaw_QLaw, Calcul
         EXPECT_TRUE(acceleration.isNear(accelerationExpected, 1e-12));
     }
 }
-
-TEST_F(OpenSpaceToolkit_Astrodynamics_Dynamics_Thruster_GuidanceLaw_QLaw, SMA_raising_with_eccentricity)
-{
-    const COE targetCOE = {
-        Length::Meters(42000.0e3),
-        0.01,
-        Angle::Degrees(0.05),
-        Angle::Degrees(0.0),
-        Angle::Degrees(0.0),
-        Angle::Degrees(0.0),
-    };
-
-    const QLaw::Parameters parameters = {
-        {
-            {COE::Element::SemiMajorAxis, 1.0},
-            {COE::Element::Eccentricity, 1.0},
-        },
-    };
-
-    const Shared<QLaw> qlaw =
-        std::make_shared<QLaw>(QLaw(targetCOE, gravitationalParameter_, parameters, finiteDifferenceSolver_));
-
-    const Mass mass = Mass::Kilograms(50.0);
-    const Composite satelliteGeometry(Cuboid(
-        {0.0, 0.0, 0.0}, {Vector3d {1.0, 0.0, 0.0}, Vector3d {0.0, 1.0, 0.0}, Vector3d {0.0, 0.0, 1.0}}, {1.0, 2.0, 3.0}
-    ));
-
-    const PropulsionSystem propulsionSystem = {
-        Scalar(1.0, PropulsionSystem::thrustSIUnit),
-        Scalar(3100.0, PropulsionSystem::specificImpulseSIUnit),
-    };
-
-    const SatelliteSystem satelliteSystem = {
-        mass,
-        satelliteGeometry,
-        Matrix3d::Identity(),
-        1.0,
-        2.1,
-        propulsionSystem,
-    };
-
-    const Shared<Thruster> thruster = std::make_shared<Thruster>(Thruster(satelliteSystem, qlaw));
-    const Shared<CentralBodyGravity> centralBodyGravity =
-        std::make_shared<CentralBodyGravity>(CentralBodyGravity(std::make_shared<Earth>(Earth::Spherical())));
-    const Shared<PositionDerivative> positionDerivative = std::make_shared<PositionDerivative>(PositionDerivative());
-    const Array<Shared<Dynamics>> dynamics = {
-        thruster,
-        centralBodyGravity,
-        positionDerivative,
-    };
-
-    const NumericalSolver numericalSolver = {
-        NumericalSolver::LogType::NoLog,
-        NumericalSolver::StepperType::RungeKutta4,
-        120.0,
-        1e-12,
-        1e-12,
-    };
-
-    const Propagator propagator = {numericalSolver, dynamics};
-
-    const COE currentCOE = {
-        Length::Meters(7000.0e3),
-        0.01,
-        Angle::Degrees(0.05),
-        Angle::Degrees(0.0),
-        Angle::Degrees(0.0),
-        Angle::Degrees(0.0),
-    };
-
-    const COE::CartesianState cartesianState = currentCOE.getCartesianState(gravitationalParameter_, Frame::GCRF());
-
-    VectorXd coordinates(7);
-    coordinates << cartesianState.first.getCoordinates(), cartesianState.second.getCoordinates(), 300.0;
-
-    const State initialState = {
-        Instant::J2000(),
-        coordinates,
-        Frame::GCRF(),
-        {CartesianPosition::Default(), CartesianVelocity::Default(), CoordinatesSubset::Mass()}
-    };
-
-    const State state = propagator.calculateStateAt(initialState, initialState.accessInstant() + Duration::Days(15.1));
-
-    const COE endCOE = COE::Cartesian({state.getPosition(), state.getVelocity()}, gravitationalParameter_);
-
-    std::cout << endCOE << std::endl;
-
-    EXPECT_TRUE(std::abs(endCOE.getSemiMajorAxis().inMeters() - targetCOE.getSemiMajorAxis().inMeters()) < 60000.0);
-}
