+++ test/OpenSpaceToolkit/Astrodynamics/Trajectory/Propagator.test.cpp
@@ -45,8 +45,11 @@
 #include <OpenSpaceToolkit/Astrodynamics/Dynamics/Thruster.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/EventCondition/InstantCondition.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/GuidanceLaw/ConstantThrust.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/GuidanceLaw/QLaw.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Solvers/FiniteDifferenceSolver.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/Trajectory/LocalOrbitalFrameDirection.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/Trajectory/Orbit/Model.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Orbit/Models/Kepler/COE.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/Trajectory/Propagator.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/Trajectory/State.hpp>
 #include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesBroker.hpp>
@@ -104,15 +107,15 @@ using ostk::physics::environment::atmospheric::earth::CSSISpaceWeather;
 using ostk::astro::eventcondition::InstantCondition;
 using ostk::astro::trajectory::State;
 using ostk::astro::trajectory::Propagator;
+using ostk::astro::trajectory::LocalOrbitalFrameFactory;
+using ostk::astro::trajectory::LocalOrbitalFrameDirection;
 using ostk::astro::trajectory::state::NumericalSolver;
 using ostk::astro::trajectory::state::CoordinatesSubset;
 using ostk::astro::trajectory::state::CoordinatesBroker;
 using ostk::astro::trajectory::state::coordinatessubsets::CartesianPosition;
 using ostk::astro::trajectory::state::coordinatessubsets::CartesianVelocity;
 using ostk::astro::trajectory::state::NumericalSolver;
-using ostk::astro::trajectory::LocalOrbitalFrameFactory;
-using ostk::astro::trajectory::LocalOrbitalFrameDirection;
-using ostk::astro::trajectory::Propagator;
+using ostk::astro::trajectory::orbit::models::kepler::COE;
 using ostk::astro::Dynamics;
 using ostk::astro::flight::system::PropulsionSystem;
 using ostk::astro::flight::system::SatelliteSystem;
@@ -122,6 +125,8 @@ using ostk::astro::dynamics::ThirdBodyGravity;
 using ostk::astro::dynamics::AtmosphericDrag;
 using ostk::astro::dynamics::Thruster;
 using ostk::astro::guidancelaw::ConstantThrust;
+using ostk::astro::guidancelaw::QLaw;
+using ostk::astro::solvers::FiniteDifferenceSolver;
 
 static const Derived::Unit GravitationalParameterSIUnit =
     Derived::Unit::GravitationalParameter(Length::Unit::Meter, ostk::physics::units::Time::Unit::Second);
@@ -2506,6 +2511,197 @@ class OpenSpaceToolkit_Astrodynamics_Trajectory_Orbit_Models_Propagator_Thruster
     Shared<Celestial> earthSpherical_ = nullptr;
     Propagator defaultPropagator_ = Propagator::Undefined();
 };
+
+TEST_F(OpenSpaceToolkit_Astrodynamics_Trajectory_Orbit_Models_Propagator, QLaw_Paper_Case_A)
+{
+    const COE targetCOE = {
+        Length::Meters(42000.0e3),
+        0.01,
+        Angle::Degrees(0.05),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+    };
+
+    const QLaw::Parameters parameters = {
+        {
+            {COE::Element::SemiMajorAxis, 1.0},
+            {COE::Element::Eccentricity, 1.0},
+        },
+    };
+
+    const Derived gravitationalParameter =
+        Derived(398600.49 * 1e9, EarthGravitationalModel::EGM2008.gravitationalParameter_.getUnit());
+
+    const Shared<QLaw> qlaw =
+        std::make_shared<QLaw>(QLaw(targetCOE, gravitationalParameter, parameters, FiniteDifferenceSolver::Default()));
+
+    const Mass mass = Mass::Kilograms(50.0);
+    const Composite satelliteGeometry(Cuboid(
+        {0.0, 0.0, 0.0}, {Vector3d {1.0, 0.0, 0.0}, Vector3d {0.0, 1.0, 0.0}, Vector3d {0.0, 0.0, 1.0}}, {1.0, 2.0, 3.0}
+    ));
+
+    const PropulsionSystem propulsionSystem = {
+        Scalar(1.0, PropulsionSystem::thrustSIUnit),
+        Scalar(3100.0, PropulsionSystem::specificImpulseSIUnit),
+    };
+
+    const SatelliteSystem satelliteSystem = {
+        mass,
+        satelliteGeometry,
+        Matrix3d::Identity(),
+        1.0,
+        2.1,
+        propulsionSystem,
+    };
+
+    const Shared<Thruster> thruster = std::make_shared<Thruster>(Thruster(satelliteSystem, qlaw));
+    const Shared<CentralBodyGravity> centralBodyGravity =
+        std::make_shared<CentralBodyGravity>(CentralBodyGravity(std::make_shared<Earth>(Earth::Spherical())));
+    const Shared<PositionDerivative> positionDerivative = std::make_shared<PositionDerivative>(PositionDerivative());
+    const Array<Shared<Dynamics>> dynamics = {
+        thruster,
+        centralBodyGravity,
+        positionDerivative,
+    };
+
+    const NumericalSolver numericalSolver = {
+        NumericalSolver::LogType::NoLog,
+        NumericalSolver::StepperType::RungeKutta4,
+        120.0,
+        1e-12,
+        1e-12,
+    };
+
+    const Propagator propagator = {numericalSolver, dynamics};
+
+    const COE currentCOE = {
+        Length::Meters(7000.0e3),
+        0.01,
+        Angle::Degrees(0.05),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+    };
+
+    const COE::CartesianState cartesianState = currentCOE.getCartesianState(gravitationalParameter, Frame::GCRF());
+
+    VectorXd coordinates(7);
+    coordinates << cartesianState.first.getCoordinates(), cartesianState.second.getCoordinates(), 300.0;
+
+    const State initialState = {
+        Instant::J2000(),
+        coordinates,
+        Frame::GCRF(),
+        {CartesianPosition::Default(), CartesianVelocity::Default(), CoordinatesSubset::Mass()}
+    };
+
+    const State state = propagator.calculateStateAt(initialState, initialState.accessInstant() + Duration::Days(15.1));
+
+    const COE endCOE = COE::Cartesian({state.getPosition(), state.getVelocity()}, gravitationalParameter);
+
+    EXPECT_TRUE(std::abs(endCOE.getSemiMajorAxis().inMeters() - targetCOE.getSemiMajorAxis().inMeters()) < 60000.0);
+}
+
+TEST_F(OpenSpaceToolkit_Astrodynamics_Trajectory_Orbit_Models_Propagator, QLaw_Paper_Case_E)
+{
+    const COE targetCOE = {
+        Length::Meters(26500.0e3),
+        0.7,
+        Angle::Degrees(116.0),
+        Angle::Degrees(180.0),
+        Angle::Degrees(270.0),
+        Angle::Degrees(0.0),
+    };
+
+    const QLaw::Parameters parameters = {
+        {
+            {COE::Element::SemiMajorAxis, 1.0},
+            {COE::Element::Eccentricity, 1.0},
+            {COE::Element::Inclination, 1.0},
+            {COE::Element::Raan, 1.0},
+            {COE::Element::Aop, 1.0},
+        },
+    };
+
+    const Derived gravitationalParameter =
+        Derived(398600.49 * 1e9, EarthGravitationalModel::EGM2008.gravitationalParameter_.getUnit());
+
+    const Shared<QLaw> qlaw =
+        std::make_shared<QLaw>(QLaw(targetCOE, gravitationalParameter, parameters, FiniteDifferenceSolver::Default()));
+
+    const Mass mass = Mass::Kilograms(50.0);
+    const Composite satelliteGeometry(Cuboid(
+        {0.0, 0.0, 0.0}, {Vector3d {1.0, 0.0, 0.0}, Vector3d {0.0, 1.0, 0.0}, Vector3d {0.0, 0.0, 1.0}}, {1.0, 2.0, 3.0}
+    ));
+
+    const PropulsionSystem propulsionSystem = {
+        Scalar(2.0, PropulsionSystem::thrustSIUnit),
+        Scalar(2000.0, PropulsionSystem::specificImpulseSIUnit),
+    };
+
+    const SatelliteSystem satelliteSystem = {
+        mass,
+        satelliteGeometry,
+        Matrix3d::Identity(),
+        1.0,
+        2.1,
+        propulsionSystem,
+    };
+
+    const Shared<Thruster> thruster = std::make_shared<Thruster>(Thruster(satelliteSystem, qlaw));
+    const Shared<CentralBodyGravity> centralBodyGravity =
+        std::make_shared<CentralBodyGravity>(CentralBodyGravity(std::make_shared<Earth>(Earth::Spherical())));
+    const Shared<PositionDerivative> positionDerivative = std::make_shared<PositionDerivative>(PositionDerivative());
+    const Array<Shared<Dynamics>> dynamics = {
+        thruster,
+        centralBodyGravity,
+        positionDerivative,
+    };
+
+    const NumericalSolver numericalSolver = {
+        NumericalSolver::LogType::NoLog,
+        NumericalSolver::StepperType::RungeKutta4,
+        120.0,
+        1e-12,
+        1e-12,
+    };
+
+    const Propagator propagator = {numericalSolver, dynamics};
+
+    const COE currentCOE = {
+        Length::Meters(24505.9e3),
+        0.725,
+        Angle::Degrees(0.06),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+        Angle::Degrees(0.0),
+    };
+
+    const COE::CartesianState cartesianState = currentCOE.getCartesianState(gravitationalParameter, Frame::GCRF());
+
+    VectorXd coordinates(7);
+    coordinates << cartesianState.first.getCoordinates(), cartesianState.second.getCoordinates(), 2000.0;
+
+    const State initialState = {
+        Instant::J2000(),
+        coordinates,
+        Frame::GCRF(),
+        {CartesianPosition::Default(), CartesianVelocity::Default(), CoordinatesSubset::Mass()}
+    };
+
+    const State state = propagator.calculateStateAt(initialState, initialState.accessInstant() + Duration::Days(81.61));
+
+    const COE endCOE = COE::Cartesian({state.getPosition(), state.getVelocity()}, gravitationalParameter);
+
+    EXPECT_TRUE(std::abs(endCOE.getSemiMajorAxis().inMeters() - targetCOE.getSemiMajorAxis().inMeters()) < 50000.0);
+    EXPECT_TRUE(std::abs(endCOE.getEccentricity() - targetCOE.getEccentricity()) < 1e-2);
+
+    // TBI: These don't close yes, have to investigate why
+    // EXPECT_TRUE(std::abs(endCOE.getInclination().inDegrees() - targetCOE.getInclination().inDegrees()) < 1.0);
+    // EXPECT_TRUE(std::abs(endCOE.getRaan().inDegrees() - targetCOE.getRaan().inDegrees()) < 1.0);
+    // EXPECT_TRUE(std::abs(endCOE.getAop().inDegrees() - targetCOE.getAop().inDegrees()) < 1.0);
+}
 
 TEST_P(
     OpenSpaceToolkit_Astrodynamics_Trajectory_Orbit_Models_Propagator_Thruster,
